Move testable code into lib/ui
